Added importable curried version of applyGroove and removed the short-circuit for builtin component props (faster to just do mixin lookup instead of that operation for each prop)
